trading/market-analysis/pyproject.toml
[project]
name = "async-trading"
version = "0.1.0"
description = "Asynchronous crypto trading analytics system"
authors = [{ name = "Settor" }]
requires-python = ">=3.12"

dependencies = [
    "pandas",
    "numpy",
    "ta",
    "python-binance",
]
